directed models reconstruct folder structures
